+++ test/libsolidity/semanticTests/variables/public_state_overridding.sol
@@ -1,0 +1,19 @@
+contract A
+{
+	function test() external virtual returns (uint256)
+	{
+		return 5;
+	}
+}
+contract X is A
+{
+	uint256 public override test;
+
+	function set() public { test = 2; }
+}
+
+
+// ----
+// test() -> 0
+// set() ->
+// test() -> 2