+++ test/libsolidity/syntaxTests/inheritance/override/function_state_variable.sol
@@ -1,2 +1,2 @@
 interface ERC20 { function x() external returns (uint); }
-contract C is ERC20 { uint public x; }
+contract C is ERC20 { uint public override x; }