+++ test/libsolidity/syntaxTests/inheritance/override/override_ambiguous.sol
@@ -1,5 +1,4 @@
 abstract contract A {
-	int public testvar;
 	function foo() internal virtual returns (uint256);
 }
 abstract contract B {
@@ -7,9 +6,7 @@ abstract contract B {
 	function test() internal virtual returns (uint256);
 }
 abstract contract X is A, B {
-	int public override testvar;
 	function test() internal override returns (uint256) {}
 }
 // ----
-// DeclarationError: (257-284): Identifier already declared.
-// TypeError: (226-343): Derived contract must override function "foo". Function with the same name and parameter types defined in two or more base classes.
+// TypeError: (205-292): Derived contract must override function "foo". Function with the same name and parameter types defined in two or more base classes.