+++ test/libsolidity/syntaxTests/inheritance/override/override_multiple.sol
@@ -13,7 +13,6 @@ abstract contract D {
 	function foo() internal virtual returns (uint256);
 }
 contract X is A, B, C, D {
-	int public override testvar;
 	function test() internal override returns (uint256) {}
 	function foo() internal override(A, B, C, D) returns (uint256) {}
 }