+++ test/libsolidity/syntaxTests/inheritance/override/override_multiple_duplicated.sol
@@ -1,5 +1,4 @@
 abstract contract A {
-	int public testvar;
 	function foo() internal virtual returns (uint256);
 	function test(uint8 _a) internal virtual returns (uint256);
 }
@@ -15,13 +14,11 @@ abstract contract D {
 	function test() internal virtual returns (uint256);
 }
 abstract contract X is A, B, C, D {
-	int public override testvar;
 	function test() internal override(B, D, D) virtual returns (uint256);
 	function foo() internal override(A, C, B, B, B, D ,D) virtual returns (uint256);
 }
 // ----
-// DeclarationError: (529-556): Identifier already declared.
-// TypeError: (599-600): Duplicate contract "D" found in override list of "test".
-// TypeError: (672-673): Duplicate contract "B" found in override list of "foo".
-// TypeError: (675-676): Duplicate contract "B" found in override list of "foo".
-// TypeError: (681-682): Duplicate contract "D" found in override list of "foo".
+// TypeError: (548-549): Duplicate contract "D" found in override list of "test".
+// TypeError: (621-622): Duplicate contract "B" found in override list of "foo".
+// TypeError: (624-625): Duplicate contract "B" found in override list of "foo".
+// TypeError: (630-631): Duplicate contract "D" found in override list of "foo".