+++ test/libsolidity/syntaxTests/inheritance/override/override_type_mismatch.sol
@@ -16,10 +16,9 @@ abstract contract X is A, B, C, D {
 	struct MyStruct { int abc; }
 	enum ENUM { F,G,H }
 
-	int public override testvar;
 	function test() internal override virtual returns (uint256);
 	function foo() internal override(MyStruct, ENUM, A, B, C, D) virtual returns (uint256);
 }
 // ----
-// TypeError: (632-640): Expected contract but got struct X.MyStruct.
-// TypeError: (642-646): Expected contract but got enum X.ENUM.
+// TypeError: (602-610): Expected contract but got struct X.MyStruct.
+// TypeError: (612-616): Expected contract but got enum X.ENUM.