+++ test/libsolidity/syntaxTests/inheritance/override/public_var_same_name_but_different_args.sol
@@ -1,0 +1,7 @@
+contract A {
+	function foo(uint) internal virtual pure returns(uint) { return 5; }
+}
+contract X is A {
+	uint public foo;
+}
+// ----