+++ test/libsolidity/syntaxTests/inheritance/override/public_vars_multiple2.sol
@@ -1,0 +1,9 @@
+contract A {
+	function foo() external virtual pure returns(uint) { return 4; }
+	function foo(uint ) external virtual pure returns(uint) { return 4; }
+	function foo(uint , uint ) external pure virtual returns(A) {  }
+}
+contract X is A {
+	uint public override foo;
+}
+// ----